fix_name/requirements.txt
pandas
tqdm
mecab-python3==0.7
gensim
python-Levenshtein
beautifulsoup4
pyclustering
openpyxl